Online dispersion calculation One more useless refactoring
Use online (one-pass) dispersion calculation, so there more
no need for vector of indexes in each cell -> no allocations.

Create Grid class for 2d grid that with all of .at() functions to access grid elements.
It was to many same functions in GridMap (and OccupancyGrid), with row col, GridCoord, const and non-const...
So I extract it all to the class, which store raw array, handle allocate/move/delete and have all of this
annoying access functions

Also, change displaying functions. Now there a few basic functions:
 - display via image
 - display via point cloud
 - display via mesh
which just get std::function that calculate color (and z for mesh).
And some functions to calculate colors:
 - z_mean
 - dispersion
 - obstacle
So now you are free to combine them in different ways